Generalise and rename the module reporter script
Instead of being specific to the RDS instance
module, this script will produce a CSV report for
any of our terraform modules, listing the 
namespaces using it, and which version they are
using.
